refactor(ios): extract EngineConfigProjector from GameSettings
